some modifications

+++ src/data/explorer.py
@@ -1,6 +1,6 @@
-from .init import curs, conn
+from .init import curs, IntegrityError
 from model.explorer import Explorer
-
+from errors import Missing, Duplicate
 
 curs.execute(
     """create table if not exists explorer(
@@ -12,20 +12,29 @@ curs.execute(
 
 
 def row_to_model(row: tuple) -> Explorer:
-    print(f"{row=}")
-    (name, description, country) = row
-    return Explorer(name=row[0], country=row[1], description=row[2])
+    name, country, description = row
+    result = Explorer(name=name, country=country, description=description)
+
+    return result
+
+
+def dict_to_model(d: dict) -> Explorer:
+    return d
 
 
 def model_to_dict(explorer: Explorer) -> dict:
-    return explorer.model_dump() if explorer else None
+    return explorer.dict() if explorer else None
 
 
 def get_one(name: str) -> Explorer:
     qry = "select * from explorer where name=:name"
     params = {"name": name}
     curs.execute(qry, params)
-    return row_to_model(curs.fetchone())
+    row = curs.fetchone()
+    if row:
+        return row_to_model(row)
+    else:
+        raise Missing(msg=f"Explorer {name} not found")
 
 
 def get_all() -> list[Explorer]:
@@ -35,36 +44,42 @@ def get_all() -> list[Explorer]:
 
 
 def create(explorer: Explorer) -> Explorer:
-    print(f"{curs=}  {conn=}")
-    print(f"DATA {explorer=}")
-    qry = """insert into explorer (name, country, description)
-             values (:name, :country, :description)"""
-    print(f"{qry=}")
+    if not explorer:
+        return None
+    qry = """insert into explorer (name, country, description) values
+             (:name, :country, :description)"""
     params = model_to_dict(explorer)
+    print(f"{qry=}")
     print(f"{params=}")
-    data_insert = curs.execute(qry, params)
-    print(f"{data_insert=}")
-    conn.commit()
-    result = get_one(explorer.name)
-    print(f"{result=}")
-    return result
+    try:
+        curs.execute(qry, params)
+    except IntegrityError:
+        raise Duplicate(msg=f"Explorer {explorer.name} already exists")
+    return get_one(explorer.name)
 
 
 def modify(name: str, explorer: Explorer) -> Explorer:
+    if not (name and explorer):
+        return None
     qry = """update explorer
-             set country=:country,
-             name=:name,
+             set name=:name,
+             country=:country,
              description=:description
              where name=:name_orig"""
     params = model_to_dict(explorer)
     params["name_orig"] = explorer.name
-    _ = curs.execute(qry, params)
-    explorer2 = get_one(explorer.name)
-    return explorer2
+    curs.execute(qry, params)
+    if curs.rowcount == 1:
+        return get_one(explorer.name)
+    else:
+        raise Missing(msg=f"Explorer {name} not found")
 
 
-def delete(explorer: Explorer) -> bool:
+def delete(name: str):
+    if not name:
+        return False
     qry = "delete from explorer where name = :name"
-    params = {"name": explorer.name}
-    res = curs.execute(qry, params)
-    return bool(res)
+    params = {"name": name}
+    curs.execute(qry, params)
+    if curs.rowcount != 1:
+        raise Missing(msg=f"Explorer {name} not found")
